Refactored Retry Analyzer.
