+++ src/test/java/io/saucelabs/portal/qa/listeners/AnnotationTransformer.java
@@ -9,6 +9,6 @@ import java.lang.reflect.Method;
 public class AnnotationTransformer implements IAnnotationTransformer {
     @Override
     public void transform(ITestAnnotation annotation, Class testClass, Constructor testConstructor, Method testMethod) {
-        annotation.setRetryAnalyzer(SauceLabsPortalRetryAnalyzer.class);
+        annotation.setRetryAnalyzer(SauceLabsPortalListener.class);
     }
 }
