+++ src/test/java/io/saucelabs/portal/qa/listeners/SauceLabsPortalListener.java
@@ -5,10 +5,7 @@ import io.saucelabs.portal.qa.commons.BaseTest;
 import lombok.extern.slf4j.Slf4j;
 import org.openqa.selenium.OutputType;
 import org.openqa.selenium.TakesScreenshot;
-import org.testng.ISuite;
-import org.testng.ISuiteListener;
-import org.testng.ITestListener;
-import org.testng.ITestResult;
+import org.testng.*;
 import org.zeroturnaround.zip.commons.FileUtils;
 
 import java.io.File;
@@ -18,9 +15,11 @@ import java.time.Instant;
 import java.util.Date;
 
 @Slf4j
-public class SauceLabsPortalListener extends BaseTest implements ITestListener, ISuiteListener {
+public class SauceLabsPortalListener extends BaseTest implements ITestListener, ISuiteListener, IRetryAnalyzer {
 
     private Instant startDate;
+    private int COUNT = 0;
+    private static final int MAX_RETRY = 3;
 
     @Override
     public void onStart(ISuite suite) {
@@ -56,18 +55,56 @@ public class SauceLabsPortalListener extends BaseTest implements ITestListener, 
         File destinationFile;
         try {
             if (testResult.isSuccess()) {
-                destinationFile = new File(SauceLabsPortalConstant.DIRECTORY + SauceLabsPortalConstant.PASS + File.separator
-                        + SauceLabsPortalConstant.PASS_PREFIX + testResult.getName() + "_" + new Date() + SauceLabsPortalConstant.IMAGE_FORMAT);
+                destinationFile = new File(SauceLabsPortalConstant.DIRECTORY +
+                        SauceLabsPortalConstant.PASS + File.separator
+                        + SauceLabsPortalConstant.PASS_PREFIX
+                        + testResult.getName() + "_" + new Date()
+                        + SauceLabsPortalConstant.IMAGE_FORMAT);
                 FileUtils.copyFile(sourceFile, destinationFile);
-                log.info("Success scenario has been captured. PASSED Screenshot has been placed in the location {}", destinationFile);
+                log.info("Success scenario has been captured. PASSED Screenshot has been placed in the location {}",
+                        destinationFile);
             } else {
-                destinationFile = new File(SauceLabsPortalConstant.DIRECTORY + SauceLabsPortalConstant.FAIL + File.separator
-                        + SauceLabsPortalConstant.FAIL_PREFIX + testResult.getName() + "_" + new Date() + SauceLabsPortalConstant.IMAGE_FORMAT);
+                destinationFile = new File(SauceLabsPortalConstant.DIRECTORY
+                        + SauceLabsPortalConstant.FAIL + File.separator
+                        + SauceLabsPortalConstant.FAIL_PREFIX
+                        + testResult.getName() + "_" + new Date()
+                        + SauceLabsPortalConstant.IMAGE_FORMAT);
                 FileUtils.copyFile(sourceFile, destinationFile);
-                log.info("Failed scenario has been captured. FAILED Screenshot has been placed in the location {}", destinationFile);
+                log.info("Failed scenario has been captured. FAILED Screenshot has been placed in the location {}",
+                        destinationFile);
             }
         } catch (IOException e) {
             log.error("Unable to capture the screenshots for Sauce Labs Portal {}", e.getMessage());
         }
+    }
+
+    @Override
+    public boolean retry(ITestResult result) {
+        if (!result.isSuccess()) {
+            if (COUNT < MAX_RETRY) {
+                log.info("Retrying test for {} time(s) for the test method {} with test status {}.",
+                        COUNT + 1, result.getName(), getTestStatusName(result.getStatus()));
+                COUNT++;
+                return true;
+            }
+        }
+        log.info("Retrying for the test method {} is exhausted.", result.getName());
+        return false;
+    }
+
+    private String getTestStatusName(int status) {
+        String testStatus = null;
+        switch (status) {
+            case 1:
+                testStatus = "SUCCESS";
+                break;
+            case 2:
+                testStatus = "FAILED";
+                break;
+            case 3:
+                testStatus = "SKIP";
+                break;
+        }
+        return testStatus;
     }
 }
